tests/requirements-test.txt
# Dépendances pour les tests FinAgent
# Installation : pip install -r tests/requirements-test.txt

# Framework de tests principal
pytest>=7.4.4,<8.0.0
pytest-asyncio>=0.23.3,<1.0.0
pytest-mock>=3.12.0,<4.0.0
pytest-cov>=4.1.0,<5.0.0
pytest-xdist>=3.5.0,<4.0.0
pytest-timeout>=2.2.0,<3.0.0
pytest-benchmark>=4.0.0,<5.0.0
pytest-randomly>=3.15.0,<4.0.0

# Coverage et rapports
coverage>=7.4.0,<8.0.0
coverage[toml]>=7.4.0,<8.0.0

# Mocking et fixtures
responses>=0.24.1,<1.0.0
httpx-mock>=0.10.0,<1.0.0
factory-boy>=3.3.0,<4.0.0
freezegun>=1.4.0,<2.0.0

# Tests de performance
pytest-benchmark>=4.0.0,<5.0.0
memory-profiler>=0.61.0,<1.0.0
psutil>=5.9.6,<6.0.0

# Tests de sécurité
bandit>=1.7.5,<2.0.0
safety>=3.0.1,<4.0.0

# Utilitaires de test
faker>=22.2.0,<23.0.0
hypothesis>=6.92.1,<7.0.0
dirty-equals>=0.7.1,<1.0.0

# Tests CLI
click-testing>=2.1,<3.0.0

# Tests database
pytest-postgresql>=5.0.0,<6.0.0
sqlalchemy-utils>=0.41.1,<1.0.0

# Utilitaires debugging
pytest-pdb>=0.2.0,<1.0.0
pytest-pudb>=0.7.0,<1.0.0

# Format et linting pour tests
ruff>=0.1.9,<1.0.0
black>=23.12.1,<24.0.0

# Documentation tests
pytest-doctestplus>=1.0.0,<2.0.0

# Tests parallèles et distribués
pytest-parallel>=0.1.1,<1.0.0

# Validation de schémas
jsonschema>=4.20.0,<5.0.0
cerberus>=1.3.4,<2.0.0